+++ src/main/java/net/ccbluex/liquidbounce/ui/client/hud/element/elements/Inventory.kt
@@ -28,7 +28,7 @@ class Inventory : Element("Inventory", 300.0, 50.0) {
     private val font by font("Font", Fonts.fontSemibold35)
     private val title by choices("Title", arrayOf("Center", "Left", "Right", "None"), "Left")
     private val titleColor = color("TitleColor", Color.WHITE) { title != "None" }
-    private val roundedRectRadius by float("Rounded-Radius", 3F, 0F..5F)
+    private val roundedRectRadius by float("RoundedRadius", 3F, 0F..5F)
 
     private val borderValue by boolean("Border", true)
     private val borderColor = color("BorderColor", Color.WHITE) { borderValue }