+++ src/main/java/net/ccbluex/liquidbounce/ui/client/hud/element/elements/Text.kt
@@ -126,44 +126,44 @@ class Text(x: Double = 10.0, y: Double = 10.0, scale: Float = 1F, side: Side = S
 
     private var displayString by text("DisplayText", "")
 
-    private val textColorMode by choices("Text-ColorMode", arrayOf("Custom", "Rainbow", "Gradient"), "Custom")
+    private val textColorMode by choices("TextColorMode", arrayOf("Custom", "Rainbow", "Gradient"), "Custom")
 
     private val colors = ColorSettingsInteger(this, "TextColor", applyMax = true) { textColorMode == "Custom" }
 
-    private val gradientTextSpeed by float("Text-Gradient-Speed", 1f, 0.5f..10f) { textColorMode == "Gradient" }
+    private val gradientTextSpeed by float("TextGradientSpeed", 1f, 0.5f..10f) { textColorMode == "Gradient" }
 
-    private val maxTextGradientColors by int("Max-Text-Gradient-Colors", 4, 1..MAX_GRADIENT_COLORS)
+    private val maxTextGradientColors by int("MaxTextGradientColors", 4, 1..MAX_GRADIENT_COLORS)
     { textColorMode == "Gradient" }
     private val textGradColors = ColorSettingsFloat.create(this, "Text-Gradient")
     { textColorMode == "Gradient" && it <= maxTextGradientColors }
 
-    private val roundedBackgroundRadius by float("RoundedBackGround-Radius", 3F, 0F..5F)
+    private val roundedBackgroundRadius by float("RoundedBackgroundRadius", 3F, 0F..5F)
 
-    private var backgroundScale by float("Background-Scale", 1F, 1F..3F)
+    private var backgroundScale by float("BackgroundScale", 1F, 1F..3F)
 
-    private val backgroundMode by choices("Background-ColorMode", arrayOf("Custom", "Rainbow", "Gradient"), "Custom")
+    private val backgroundMode by choices("BackgroundColorMode", arrayOf("Custom", "Rainbow", "Gradient"), "Custom")
 
     private val bgColors = ColorSettingsInteger(this, "BackgroundColor")
     { backgroundMode == "Custom" }.with(a = 0)
 
-    private val gradientBackgroundSpeed by float("Background-Gradient-Speed", 1f, 0.5f..10f)
+    private val gradientBackgroundSpeed by float("BackgroundGradientSpeed", 1f, 0.5f..10f)
     { backgroundMode == "Gradient" }
 
-    private val maxBackgroundGradientColors by int("Max-Background-Gradient-Colors", 4, 1..MAX_GRADIENT_COLORS)
+    private val maxBackgroundGradientColors by int("MaxBackgroundGradientColors", 4, 1..MAX_GRADIENT_COLORS)
     { backgroundMode == "Gradient" }
-    private val bgGradColors = ColorSettingsFloat.create(this, "Background-Gradient")
+    private val bgGradColors = ColorSettingsFloat.create(this, "BackgroundGradient")
     { backgroundMode == "Gradient" && it <= maxBackgroundGradientColors }
 
-    private val backgroundBorder by float("BackgroundBorder-Width", 0.5F, 0.5F..5F)
+    private val backgroundBorder by float("BackgroundBorderWidth", 0.5F, 0.5F..5F)
 
     private val bgBorderColors = ColorSettingsInteger(this, "BackgroundBorderColor").with(a = 0)
 
     private fun isColorModeUsed(value: String) = textColorMode == value || backgroundMode == value
 
-    private val rainbowX by float("Rainbow-X", -1000F, -2000F..2000F) { isColorModeUsed("Rainbow") }
-    private val rainbowY by float("Rainbow-Y", -1000F, -2000F..2000F) { isColorModeUsed("Rainbow") }
-    private val gradientX by float("Gradient-X", -500F, -2000F..2000F) { isColorModeUsed("Gradient") }
-    private val gradientY by float("Gradient-Y", -1500F, -2000F..2000F) { isColorModeUsed("Gradient") }
+    private val rainbowX by float("RainbowX", -1000F, -2000F..2000F) { isColorModeUsed("Rainbow") }
+    private val rainbowY by float("RainbowY", -1000F, -2000F..2000F) { isColorModeUsed("Rainbow") }
+    private val gradientX by float("GradientX", -500F, -2000F..2000F) { isColorModeUsed("Gradient") }
+    private val gradientY by float("GradientY", -1500F, -2000F..2000F) { isColorModeUsed("Gradient") }
 
     private var shadow by boolean("Shadow", true)
     private val font = font("Font", Fonts.fontSemibold40)